+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - trapFocus: Fixed aria-hidden not removing for some of the event handlers combinations
 - Carousel: Fixed scroll snapping scrolling the carousel back
 - TextField: Rounded flag
+- Flyout, Popover, DropdownMenu, Tooltip: contentShift, contentGap
 
 Documentation